// src/order_codes/orderCodes.types.ts
export interface datesOrder {
    id_order: number;
    code: string;
  }

  export interface NewOrderCode {
    id_order: number;
  }

export interface updateInfo {
    id_user: number;
    id_compani: string;
  }
  
  export interface updateInfoSec {
    id_user: string;
    id_compani: string;
  }
  